[Event Hubs] Move T1 migration guide snippets
The focus of these changes is to move the migration snippets demonstrating
use of `Microsoft.Azure.EventHubs` to the `Azure.Messaging.EventHubs` test
projects.  This consolidates the build pipeline used for validating the
snippets and allows for removal of the T1 code once the library is formally
retired.
